updated floats to bigdecimal. #2681

+++ j-lawyer-client/src/com/jdimension/jlawyer/client/editors/files/CaseAccountEntryPanel.java
@@ -671,6 +671,7 @@ import com.jdimension.jlawyer.persistence.Invoice;
 import java.awt.Component;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
+import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
@@ -1164,12 +1165,12 @@ public class CaseAccountEntryPanel extends javax.swing.JPanel {
         this.entry.setEntryDate(this.df.parse(this.txtDate.getText()));
         this.entry.setContact(this.recipientAddress);
         this.entry.setDescription(this.taDescription.getText());
-        this.entry.setEarnings(((Number) txtEarnings.getValue()).floatValue());
-        this.entry.setSpendings(((Number) txtSpendings.getValue()).floatValue());
-        this.entry.setEscrowIn(((Number) txtEscrowIn.getValue()).floatValue());
-        this.entry.setEscrowOut(((Number) txtEscrowOut.getValue()).floatValue());
-        this.entry.setExpendituresIn(((Number) txtExpendituresIn.getValue()).floatValue());
-        this.entry.setExpendituresOut(((Number) txtExpendituresOut.getValue()).floatValue());
+        this.entry.setEarnings(BigDecimal.valueOf(((Number) txtEarnings.getValue()).floatValue()));
+        this.entry.setSpendings(BigDecimal.valueOf(((Number) txtSpendings.getValue()).floatValue()));
+        this.entry.setEscrowIn(BigDecimal.valueOf(((Number) txtEscrowIn.getValue()).floatValue()));
+        this.entry.setEscrowOut(BigDecimal.valueOf(((Number) txtEscrowOut.getValue()).floatValue()));
+        this.entry.setExpendituresIn(BigDecimal.valueOf(((Number) txtExpendituresIn.getValue()).floatValue()));
+        this.entry.setExpendituresOut(BigDecimal.valueOf(((Number) txtExpendituresOut.getValue()).floatValue()));
 
         if (this.cmbInvoice.getSelectedItem() instanceof Invoice) {
             this.entry.setInvoice((Invoice) this.cmbInvoice.getSelectedItem());
